drop unsupported cucumber versions
