--- allure-plugin/src/it/cucumber-jvm/build.gradle
@@ -1,29 +1,0 @@
-plugins {
-    id 'java'
-    id 'io.qameta.allure'
-}
-
-allure {
-    aspectjweaver = true
-
-    useCucumberJVM {
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'commons-io:commons-io:2.5'
-    implementation 'info.cukes:gherkin:2.12.2'
-    implementation 'info.cukes:cucumber-core:1.2.5'
-    implementation 'info.cukes:cucumber-java:1.2.5'
-    implementation 'info.cukes:cucumber-junit:1.2.5'
-
-    testImplementation 'junit:junit:4.12'
-}
-
-test {
-    useJUnit()
-}
