--- allure-plugin/src/it/cucumber-jvm/src/test/java/tests/CucumberJvmTest.java
@@ -1,12 +1,0 @@
-package tests;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        features = {"src/test/resources/features"},
-        plugin = {"io.qameta.allure.cucumberjvm.AllureCucumberJvm"})
-public class CucumberJvmTest {
-}
