--- allure-plugin/src/it/cucumber2-jvm/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-    id 'java'
-    id 'io.qameta.allure'
-}
-
-allure {
-    useCucumber2JVM {
-    }
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'commons-io:commons-io:2.5'
-    implementation 'info.cukes:gherkin:2.12.2'
-    implementation 'io.cucumber:cucumber-core:2.3.1'
-    implementation 'io.cucumber:cucumber-java:2.3.1'
-    implementation 'io.cucumber:cucumber-junit:2.3.1'
-
-    testImplementation 'junit:junit:4.12'
-}
-
-test {
-    useJUnit()
-}
