--- allure-plugin/src/it/cucumber2-jvm/src/test/resources/features/test.feature
@@ -1,5 +1,0 @@
-@tag
-Feature: Test one
-
-  Scenario: Outline
-    Given step
