+++ allure-plugin/src/it/cucumber7-jvm/build.gradle
@@ -4,13 +4,10 @@ plugins {
 }
 
 allure {
-    version = '2.8.0'
-    aspectjweaver = true
-
     adapter {
         frameworks {
-            junit4 {
-                version = "2.9.0"
+            cucumber7jvm {
+                version = "2.28.0"
             }
         }
     }
@@ -21,6 +18,10 @@ repositories {
 }
 
 dependencies {
+    testImplementation platform('io.cucumber:cucumber-bom:7.13.0')
+    testImplementation 'io.cucumber:cucumber-junit'
+    testImplementation 'io.cucumber:cucumber-java'
+
     testImplementation 'junit:junit:4.12'
 }
