+++ allure-plugin/src/it/cucumber7-jvm/src/test/java/tests/Cucumber7JvmTest.java
@@ -1,0 +1,12 @@
+package tests;
+
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(
+        features = {"src/test/resources/features"},
+        plugin = {"io.qameta.allure.cucumber7jvm.AllureCucumber7Jvm"})
+public class Cucumber7JvmTest {
+}
